requirements.txt
cocotb-test
numpy
sympy